Fix MBL zigzag calls

+++ Marlin/Marlin_main.cpp
@@ -2955,7 +2955,7 @@ inline void gcode_G28() {
       return;
     }
 
-    int ix, iy;
+    int8_t ix, iy;
     float z;
 
     switch (state) {
